requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-migrate
flask-limiter
pymysql
requests
python-dotenv
python-docx
pypdf
openpyxl
chardet
gunicorn
werkzeug
pytest
